Move sharkiq coordinator to separate module
